fix(vant): bundle exclude vue-router
