+++ packages/vue-vant/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@cphayim-enc/vue-vant",
-  "version": "0.1.1",
+  "version": "0.1.6",
   "description": "",
   "keywords": [],
   "author": "Cphayim <i@cphayim.me>",
